requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
pyodbc==5.0.1
sqlalchemy==2.0.23

# Configuration and environment
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Data quality
great-expectations==0.18.8

# Logging and monitoring
loguru==0.7.2

# Date/time utilities
python-dateutil==2.8.2

# Excel support (for synthetic data generation)
openpyxl==3.1.2

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Type hints
types-python-dateutil==2.8.19
types-PyYAML==6.0.12
